Will test some setting tomorrow
