+++ build.gradle
@@ -1,10 +1,11 @@
+import nu.studer.gradle.jooq.JooqEdition
+
 plugins {
     id 'java'
     id "com.diffplug.spotless" version "6.6.1"
     id "application"
     id "nu.studer.jooq" version "6.0.1"
 }
-
 
 group 'io.github.org.programming'
 version '1.0-SNAPSHOT'
@@ -19,6 +20,7 @@ sourceCompatibility = 17
 targetCompatibility = 17
 
 dependencies {
+    implementation rootProject
     implementation group: 'net.dv8tion', name: 'JDA', version: '5.0.0-alpha.12'
     //Logger
     implementation group: 'ch.qos.logback', name: 'logback-classic', version: '1.2.11'
@@ -31,11 +33,11 @@ dependencies {
     implementation group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.13.3'
     //database
     implementation group: 'org.mariadb.jdbc', name: 'mariadb-java-client', version: '3.0.5'
-    implementation group: 'org.jooq', name: 'jooq', version: '3.15.5'
     implementation group: 'com.zaxxer', name: 'HikariCP', version: '5.0.0'
-    //jooq dependency
-    jooqGenerator 'org.jooq:jooq-meta:3.16.6'
-    jooqGenerator 'org.jooq:jooq-codegen:3.16.6'
+    //jooq
+    implementation group: 'org.jooq', name: 'jooq', version: '3.15.5'
+    implementation group: 'org.jooq', name: 'jooq-codegen', version: '3.16.6'
+    implementation group: 'org.jooq', name: 'jooq-meta', version: '3.16.6'
     //test
     testImplementation 'org.junit.jupiter:junit-jupiter:5.8.2'
 }
@@ -132,6 +134,44 @@ task downloadDependencies {
         }
         logger.info("  ---    Dependencies copied.  ---")
         logger.info("===== Dependencies downloaded ======")
+    }
+}
+
+jooq {
+    version = '3.15.2'
+    edition = JooqEdition.OSS
+    configurations {
+        main {
+            generateSchemaSourceOnCompilation = false
+            generationTool {
+                logging = org.jooq.meta.jaxb.Logging.WARN
+                jdbc {
+                    driver = 'org.mariadb.jdbc.Driver'
+                    //ُTODO: Make this configurable
+                    url = System.getenv("JDBC_URL")
+                    user = System.getenv("DATA_BASE_USER_NAME")
+                    password = System.getenv("DATA_BASE_PASSWORD")
+                }
+                generator {
+                    name = 'org.jooq.codegen.JavaGenerator'
+                    database {
+                        name = 'org.jooq.meta.sqlite.SQLiteDatabase'
+                        includes = '.*'
+                        excludes = ''
+                    }
+                    generate {
+                        deprecated = false
+                        records = true
+                        fluentSetters = true
+                    }
+                    target {
+                        packageName = 'io.github.org.programming.db.generated'
+                        directory = "build/generated/java"
+                    }
+                    strategy.name = "org.jooq.codegen.DefaultGeneratorStrategy"
+                }
+            }
+        }
     }
 }
